+++ cli/cli/helpers/enclave_status_from_container_status_retriever/enclave_status_from_container_status_retriever.go
@@ -2,6 +2,7 @@ package enclave_status_from_container_status_retriever
 
 import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager/types"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/container_status_calculator"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/enclave_statuses"
 	"github.com/palantir/stacktrace"
 )
@@ -11,14 +12,19 @@ import (
 func GetEnclaveStatus(containerStates []*types.Container) (enclave_statuses.EnclaveStatus, error) {
 	result := enclave_statuses.Stopped
 	for _, containerState := range containerStates {
+		containerName := containerState.GetName()
 		containerStatus := containerState.GetStatus()
-		switch containerStatus {
-		case types.Running, types.Restarting:
+		isRunning, err := container_status_calculator.IsContainerRunning(containerStatus)
+		if err != nil {
+			return "", stacktrace.Propagate(
+				err,
+				"An error occurred if container '%v' with status '%v' is running",
+				containerName,
+				containerStatus,
+			)
+		}
+		if isRunning {
 			result = enclave_statuses.Running
-		case types.Paused, types.Removing, types.Dead, types.Created, types.Exited:
-			continue
-		default:
-			return "", stacktrace.NewError("Unrecognized container status '%v'; this is a bug in Kurtosis", containerStatus)
 		}
 	}
 	return result, nil